# packages/pth-console/operator-console-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PTL Operator Console API",
    "version": "1.0.0",
    "description": "Loopback-only Console HTTP API. See docs/pth/console-protocol.md for security rules."
  },
  "servers": [{ "url": "http://127.0.0.1:3194" }],
  "paths": {
...
    "/api/v1/session/bootstrap": {
      "post": {
        "summary": "Exchange one-time fragment token for session cookie and CSRF token",
        "operationId": "bootstrapSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": { "token": { "type": "string", "pattern": "^[0-9a-f]{64}$" } }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SessionBootstrap" },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" }
        }
      }
    },
    "/api/v1/session": {
      "get": {
        "summary": "Current session state",
        "operationId": "getSession",
        "responses": {
          "200": { "$ref": "#/components/responses/SessionInfo" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" }
        }
      }
    },
...
    "/api/v1/work/actions": {
      "get": {
        "summary": "Registered native work actions",
        "operationId": "getWorkActions",
        "responses": {
          "200": {
            "description": "action list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["actions", "tenant", "space"],
                  "properties": {
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["mode", "action", "nativeKind", "descriptor"],
                        "properties": {
                          "mode": { "type": "string", "enum": ["run", "intake", "optimize"] },
                          "action": { "type": "string" },
                          "nativeKind": { "type": "string" },
                          "descriptor": {
                            "type": "object",
                            "required": ["title", "fields"],
                            "properties": {
                              "title": { "type": "string" },
                              "description": { "type": "string" },
                              "fields": { "type": "array", "items": { "$ref": "#/components/schemas/FormField" } }
                            }
                          }
                        }
                      }
                    },
                    "tenant": { "type": "string" },
                    "space": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error401" },
          "503": { "$ref": "#/components/responses/Error503" }
        }
      }
    },
    "/api/v1/work/preview": {
      "post": {
        "summary": "Preview a native work action",
        "operationId": "previewWork",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mode", "action", "input"],
                "properties": {
                  "mode": { "type": "string", "enum": ["run", "intake", "optimize"] },
                  "action": { "type": "string" },
                  "input": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "preview",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["preview", "tenant", "space"],
                  "properties": {
                    "preview": {
                      "type": "object",
                      "required": ["previewId", "previewDigest", "mode", "action", "normalizedInput", "summary", "impact", "nativeTarget", "expiresAt", "confirmation"],
                      "properties": {
                        "previewId": { "type": "string" },
                        "previewDigest": { "type": "string", "pattern": "^sha256:[0-9a-f]{64}$" },
                        "mode": { "type": "string" },
                        "action": { "type": "string" },
                        "summary": { "type": "array", "items": { "type": "string" } },
                        "impact": {
                          "type": "object",
                          "required": ["scope", "reversible", "risk"],
                          "properties": {
                            "scope": { "type": "string" },
                            "reversible": { "type": "boolean" },
                            "risk": { "type": "string" }
                          }
                        },
                        "nativeTarget": { "type": "string" },
                        "expiresAt": { "type": "string" },
                        "confirmation": { "type": "string", "const": "required" }
                      }
                    },
                    "tenant": { "type": "string" },
                    "space": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "409": { "$ref": "#/components/responses/Error409" },
          "503": { "$ref": "#/components/responses/Error503" }
        }
      }
    },
    "/api/v1/work/submit": {
      "post": {
        "summary": "Submit a previewed work action (idempotent)",
        "operationId": "submitWork",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["previewId", "previewDigest", "idempotencyKey"],
                "properties": {
                  "previewId": { "type": "string" },
                  "previewDigest": { "type": "string", "pattern": "^sha256:[0-9a-f]{64}$" },
                  "idempotencyKey": { "type": "string", "minLength": 1, "maxLength": 128 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "native work ref",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ref"],
                  "properties": {
                    "ref": {
                      "type": "object",
                      "required": ["mode", "kind", "id", "tenantId", "submittedAt"],
                      "properties": {
                        "mode": { "type": "string" },
                        "kind": { "type": "string" },
                        "id": { "type": "string" },
                        "tenantId": { "type": "string" },
                        "submittedAt": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "409": { "$ref": "#/components/responses/Error409" },
          "410": { "$ref": "#/components/responses/Error410" },
          "502": { "$ref": "#/components/responses/Error502" }
        }
      }
    },
    "/api/v1/work/evaluate": {
      "post": {
        "summary": "Evaluate native work acceptance",
        "operationId": "evaluateWork",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mode", "kind", "id"],
                "properties": {
                  "mode": { "type": "string" },
                  "kind": { "type": "string" },
                  "id": { "type": "string" },
                  "submittedAt": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "acceptance projection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["acceptance"],
                  "properties": {
                    "acceptance": {
                      "type": "object",
                      "required": ["accepted"],
                      "properties": {
                        "accepted": { "type": "boolean" },
                        "evidence": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FormField": {
        "type": "object",
        "required": ["name", "type", "required"],
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["string", "number", "boolean", "enum", "object", "array"] },
          "required": { "type": "boolean" },
          "description": { "type": "string" },
          "schema": { "type": "object" }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9_-]{0,63}$" },
              "message": { "type": "string" },
              "requestId": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
      "SessionBootstrap": {
        "description": "session created",
        "content": { "application/json": { "schema": { "type": "object", "required": ["ok", "csrfToken", "operatorPrincipalId", "expiresAt"], "properties": { "ok": { "type": "boolean" }, "csrfToken": { "type": "string" }, "operatorPrincipalId": { "type": "string" }, "expiresAt": { "type": "string" } } } } }
      },
      "SessionInfo": {
        "description": "session info",
        "content": { "application/json": { "schema": { "type": "object", "required": ["ok", "operatorPrincipalId", "expiresAt"], "properties": { "ok": { "type": "boolean" }, "operatorPrincipalId": { "type": "string" }, "expiresAt": { "type": "string" } } } } }
      },
      "Error400": { "description": "bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Error401": { "description": "unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Error403": { "description": "forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Error409": { "description": "conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Error410": { "description": "gone", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Error502": { "description": "upstream failure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Error503": { "description": "unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}